refactor(): name changes to reflect purposes better - parsers are now deserializers - writers are now serializers
