--- packages/parsers/json-parser/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "@jscad/json-parser",
-  "version": "0.0.1",
-  "description": "Json parser for jscad project",
-  "repository": "https://github.com/jscad/io",
-  "main": "index.js",
-  "scripts": {
-    "test": "ava './**/*.test.js' --verbose --timeout 10000",
-    "release-patch": "git checkout master && npm version patch && git commit -a -m 'chore(dist): built dist/'; git push origin master --tags ",
-    "release-minor": "git checkout master && npm version minor && git commit -a -m 'chore(dist): built dist/'; git push origin master --tags ",
-    "release-major": "git checkout master && npm version major && git commit -a -m 'chore(dist): built dist/'; git push origin master --tags "
-  },
-  "contributors": [
-    {
-      "name": "Rene K. Mueller",
-      "url": "http://renekmueller.com"
-    },
-    {
-      "name": "z3dev",
-      "url": "http://www.z3d.jp"
-    },
-    {
-      "name": "Mark 'kaosat-dev' Moissette",
-      "url": "http://kaosat.net"
-    }
-  ],
-  "keywords": [
-    "openjscad",
-    "jscad",
-    "csg",
-    "parser",
-    "json"
-  ],
-  "license": "MIT",
-  "dependencies": {
-  },
-  "devDependencies": {
-  }
-}
