+++ packages/serializers/svg-serializer/README.md
